Fix packages sort algorithm
We want to compare resource group first, then fallback to resource
version if the group is the same. The previous version of the sort
algorithm did not work as expected in some situations, by comparing
resources version even though their group does not match.

Only the case where groupA < groupB was considered for group comparison,
leading to a version comparison if groupA > groupB.

This commit fixes it by making sure we evaluate version comparison only
if groups are the same.

It is verified by a unit test that does not pass with the previous
version of the sort algorithm.